third_party/WebKit/LayoutTests/http/tests/misc/form-target-iframe-inline-location-change.html

<html>
<body>
<script>
if (window.testRunner) {
    testRunner.dumpResourceLoadCallbacks();
    testRunner.dumpAsText();
    testRunner.waitUntilDone();
}
</script>
<iframe id="theframe" name="theframe"></iframe>
<form action="resources/slow-frame-with-image.php" id="theform" method="post" name="theform" target="theframe"></form>
<script>
document.forms['theform'].submit();
location.href = "resources/success-notify-done.html";
</script>
</body>
</html>
